+++ cnccode/data_classes/DBEHeader.inc.php
@@ -91,6 +91,9 @@ class DBEHeader extends DBEntity
     const autoCriticalP3Hours = "autoCriticalP3Hours";
     const sevenDayerTarget = "sevenDayerTarget";
     const closureReminderDays = "closureReminderDays";
+    const solarwindsPartnerName = "solarwindsPartnerName";
+    const solarwindsUsername = "solarwindsUsername";
+    const solarwindsPassword = "solarwindsPassword";
 
     /**
      * calls constructor()
@@ -564,6 +567,9 @@ class DBEHeader extends DBEntity
         $this->addColumn(self::autoCriticalP2Hours, DA_FLOAT, DA_NOT_NULL);
         $this->addColumn(self::autoCriticalP3Hours, DA_FLOAT, DA_NOT_NULL);
         $this->addColumn(self::closureReminderDays, DA_INTEGER, DA_NOT_NULL);
+        $this->addColumn(self::solarwindsPartnerName, DA_TEXT, DA_NOT_NULL);
+        $this->addColumn(self::solarwindsUsername, DA_TEXT, DA_NOT_NULL);
+        $this->addColumn(self::solarwindsPassword, DA_TEXT, DA_NOT_NULL);
 
         $this->setPK(0);
         $this->setAddColumnsOff();